Allow new lines when pulling children
When an XML document has new lines, it contains children that are "blank" text
nodes. These nodes contain a combination of new lines and whitespace. If one of
these nodes is present as the first child, it assumes that the children contain
text.

Instead, ignore any nodes that are whitespace only.